Mode csr edge creation task and event to separate files
