+++ src/core/operator/physical_path_finding_operator.cpp
@@ -12,6 +12,7 @@
 #include "duckdb/parallel/thread_context.hpp"
 #include "duckpgq/core/utils/duckpgq_barrier.hpp"
 #include <algorithm>
+#include <duckpgq/core/operator/event/csr_edge_creation_event.hpp>
 #include <duckpgq/core/operator/event/shortest_path_event.hpp>
 #include <thread>
 
@@ -266,84 +267,6 @@ PhysicalPathFinding::Combine(ExecutionContext &context,
 //===--------------------------------------------------------------------===//
 // Finalize
 //===--------------------------------------------------------------------===//
-
-class PhysicalCSREdgeCreationTask : public ExecutorTask {
-public:
-  PhysicalCSREdgeCreationTask(shared_ptr<Event> event_p, ClientContext &context,
-                              PathFindingGlobalState &state)
-      : ExecutorTask(context, std::move(event_p)), context(context),
-        state(state) {}
-
-  TaskExecutionResult ExecuteTask(TaskExecutionMode mode) override {
-    auto &global_inputs = state.global_inputs;
-    auto &global_csr = state.global_csr;
-    auto &scan_state = state.scan_state;
-
-    DataChunk input;
-    global_inputs->InitializeScanChunk(input);
-    auto result = Vector(LogicalTypeId::BIGINT);
-    while (true) {
-      {
-        lock_guard<mutex> lock(global_csr->csr_lock);
-        if (!global_inputs->Scan(scan_state, input)) {
-          break;
-        }
-      }
-      if (!global_csr->initialized_e) {
-        const auto e_size = input.data[7].GetValue(0).GetValue<int64_t>();
-        global_csr->InitializeEdge(e_size);
-      }
-      TernaryExecutor::Execute<int64_t, int64_t, int64_t, int32_t>(
-          input.data[6], input.data[4], input.data[2], result, input.size(),
-          [&](int64_t src, int64_t dst, int64_t edge_id) {
-            const auto pos = ++global_csr->v[src + 1];
-            global_csr->e[static_cast<int64_t>(pos) - 1] = dst;
-            global_csr->edge_ids[static_cast<int64_t>(pos) - 1] = edge_id;
-            return 1;
-          });
-    }
-    event->FinishTask();
-    global_csr->Print();
-    return TaskExecutionResult::TASK_FINISHED;
-  }
-
-private:
-  ClientContext &context;
-  PathFindingGlobalState &state;
-};
-
-class CSREdgeCreationEvent : public BasePipelineEvent {
-public:
-  CSREdgeCreationEvent(PathFindingGlobalState &gstate_p, Pipeline &pipeline_p)
-      : BasePipelineEvent(pipeline_p), gstate(gstate_p) {}
-
-  PathFindingGlobalState &gstate;
-
-  void Schedule() override {
-    auto &context = pipeline->GetClientContext();
-    auto &ts = TaskScheduler::GetScheduler(context);
-    idx_t num_threads = ts.NumberOfThreads();
-    auto &scan_state = gstate.scan_state;
-    auto &global_inputs = gstate.global_inputs;
-
-    global_inputs->InitializeScan(scan_state);
-
-    vector<shared_ptr<Task>> tasks;
-    for (idx_t tnum = 0; tnum < num_threads; tnum++) {
-      tasks.push_back(make_uniq<PhysicalCSREdgeCreationTask>(shared_from_this(),
-                                                             context, gstate));
-    }
-    SetTasks(std::move(tasks));
-  }
-
-  void FinishEvent() override {
-    auto &gstate = this->gstate;
-    auto &global_csr = gstate.global_csr;
-    global_csr->is_ready = true;
-  }
-};
-
-
 
 SinkFinalizeType
 PhysicalPathFinding::Finalize(Pipeline &pipeline, Event &event,
