+++ src/core/operator/task/CMakeLists.txt
@@ -1,4 +1,5 @@
 set(EXTENSION_SOURCES
+        ${CMAKE_CURRENT_SOURCE_DIR}/csr_edge_creation_task.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/iterative_length_task.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/shortest_path_task.cpp
         ${EXTENSION_SOURCES}
